pyproject.toml
[tool.poetry]
name = "markup"
version = "0.1.0"
description = "markdown 解析器"
authors = ["ayuge <ayugesheng@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.3.2"
pywebview = "^4.3.3"
toml = "0.10.2"

[tool.poetry.group.dev.dependencies]
markdown = "^3.4.4"
commonmark = "^0.9.1"
PyInstaller = [
    {version = "6.1.0", python = ">=3.8,<3.13"},
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
